src/analytics-ai/data-pipeline-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
aiokafka==0.10.0
pydantic==2.7.3
pydantic-settings==2.3.4
cassandra-driver==3.29.1
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0